SC: Validate sync call entry point

+++ libraries/chain/include/eosio/chain/webassembly/interface.hpp
@@ -694,10 +694,11 @@ namespace webassembly {
           *                the sync call. LSB bit indicates read-only. All other bits
           *                are reserved to be 0.
           * @param data - the data of the sync call, which may include function name, arguments and other information.
-          * @return the number of bytes of the return value of the call (to be used by next get_call_return_value) .
+          * @return -1 if sync_call entry point does not exist or has invalid signature
+          * @return the number of bytes of the return value of the call (to be used by next get_call_return_value) if succeeded
           *
          */
-         uint32_t call(name receiver, uint64_t flags, span<const char> data);
+         int64_t call(name receiver, uint64_t flags, span<const char> data);
 
          /**
           * Copies the last sync call return value to `memory` up to the length of `memory`.
